Introduce SelectionRenderer
